bindings: lowercase an ASCII character in a locale-proof manner
Calling tolower() on the 'I' (of "M-I") in a Turkish locale will not
return 'i' but 'I' instead.  This is because the lowercase of 'I' is
'ı' (a dotless i), which does not fit in an 8-bit character type, so
tolower() will return the 'I' unchanged.

This fixes https://savannah.gnu.org/bugs/?67092.

Buglet existed since before version 2.2.0.