new feature: save and restore anchors when the first line has an anchor
When an anchor is placed on the first line in a buffer and --positionlog
is active, nano will now save the positions of the anchors when the file
is closed, and will restore the anchors when the file is reopened.

(The positions are stored in the file-positions history file, along
with the cursor position.)

This fulfills https://savannah.gnu.org/bugs/?62269.
Requested-by: Shi Yanling <sylphenix@126.com>
Requested-by: Sébastien Desreux <seb@h-k.fr>

Sponsored-by: Sébastien Desreux <https://doc-solus.fr>