feedback: change the appearance of anchors when they will be saved
When an anchor is present on the first line of the buffer, make any
anchors appear as little blocks (￭) instead of the usual diamonds (⬥).

This different appearance lets the user know whether a placed anchor
anywhere in the file will be saved or not -- when line numbers are
active.  (In an 8-bit locale an = replaces the usual +).