history: reverse the order in the file-positions list, to latest first
Putting the record for the last-edited file at the top of the list
ensures that (in case of a duplicate entry) the newest entry will
be encountered first and thus reused.  It also makes searching the
list faster, because the files that the user is likely to soon edit
again will be near the top of the list instead of near the end.

This fixes https://savannah.gnu.org/bugs/?67120.

Problem existed since version 2.3.0, since position logging was
introduced, but was aggravated by the recent anchor-saving code.