history: do not forget anchors when line number is given on command line
When a line number is given on the command line, the position-history
file needs to be read first, because it might contain saved anchors.

(Before placing the cursor on the given line, or searching for a given
string, the current position now needs to be set back to top-of-file,
as otherwise the saved column position is likely to be inherited, or
the search will start from the inherited position.)

This fixes https://savannah.gnu.org/bugs/?67272.

Bug existed since version 8.5, commit 60e2f2ba,
since the saving of anchors was introduced.