browser: let a refresh also reread the file list, like it used to
Until two commits ago, a full refresh (^L) would first refresh the
screen and then reread the file list and redo the screen again.
Now it will just reread the list and redo the screen, skipping the
redundant refresh.  The latter is done only for the tiny version.

Also rename a label, to be more precise.